:sparkles: New Commenting.DocblockFormat sniff
New sniff to verify the formatting of a docblock comment.

* A docblock comment should not be empty.
* Single-line docblocks are allowed by default. This can be turned off using the `allow_short` property.
* Single-line docblocks should have one space after the opener and before the closer.
* For multi-line docblocks:
    - The docblock opener and closer should each be the only content on their own line for multi-line docblocks.
    - The first content of the docblock should start on the line below the opener.
    - The last content of the docblock should be on the line above the closer.
    - Each line in the docblock should start with a star.
    - Stars should be aligned with the first star in the docblock opener.
    - There should be a minimum of one space after each star (unless it is an empty comment line).
    - If the docblock has tags, there should be an empty line before the first tag.
    - There should be no consecutive blank lines in a docblock.

Ref: https://make.wordpress.org/core/handbook/best-practices/inline-documentation-standards/php/

This sniff doesn't concern itself with the indentation of the docblock opener, as that is the concern of the `ScopeIndent` sniff.
Once the `ScopeIndent` sniff has done it's work, however, this sniff will ensure that the indentation of subsequent lines of the docblock is aligned when compared to the indentation of the docblock opener on the first line.

The sniff works independently of tabs vs spaces and will base the resulting indentation on the indentation found on the first line. Tabs vs spaces is the concern of another sniff.

Includes extensive unit tests. File `1` is tab-based. File `2` is the same, but space based.
File `3` tests a couple of very specific cases which are tokenized differently between PHPCS 2.x and PHPCS 3.x, which is why the fixer for these cases is not tested as the output would not match for one of the two supported versions.