Sniff: Add $phpcsCommentTokens array for PHPCS cross-version compatibility
PHPCS 3.2.0 introduces a new flexible PHPCS native whitelist comment syntax.
The whitelist comments available are each assigned a new PHPCS native token.

As the PHPCS constants for these tokens will not be available in older PHPCS versions, to allow for these comments, sniffs will need to check the `type` of a token, rather than the `code` and will need to use the new property as introduced in this commit.

The new property can be removed when the WPCS minimum requirement has gone up to PHPCS 3.2.3 as in that version PHPCS introduces a static array in the `Tokens` class for the same group of tokens.

Code using the WPCS version of the array will need to be adjusted at that time.