PrecisionAlignment: examine PHPCS annotations
Treat the new `// phpcs:` comments in the same way as "ordinary" `T_COMMENT` tokens.

Note: Errors will only show up when using `--ignore-annotations` as errors for lines which only contain PHPCS annotations are suppressed by default.

Includes unit test in a separate file to allow for testing with `--ignore-annotations`.

N.B.: The changes in this sniff will only work in combination with PHPCS 3.2.3+ due to upstream bugs in the tokenizing of the new PHPCS annotations in earlier PHPCS versions.