FileInclude: add "file(s) allowed to be included/required" feature
This PR was inspired by PR 104.

The `FileInclude` sniff already contained a whitelist for files from which `include`/`require` calls are expected and allowed.

However, files are often enough included/required from arbitrarily named files in a theme and there are certain typical library files, such as the TGMPA file, which are always fine to be included.

This PR adds the functionality needed to also have a whitelist of files which are allowed to be included from any file.

The sniff will only test the *last* text string in an include/require statement and will only accept a file as "allowed to be included" if the file name is either the complete string or at the end of the text string preceded by a `/`.

Includes unit test.

Other notes:
* Minor efficiency fix. Once it has been established that a file is on the "whitelist for files from which `include`/`require` calls are expected", this sniff will not be called again for that file.
* The "whitelist for files from which `include`/`require` calls are expected" functionality was not yet unit tested. A test has been added for this.
* Left the TGMPA file in the "whitelist for files from which `include`/`require` calls are expected" as well as it includes WP core files if/when needed.
* Changed the errorcode from `FileIncludeFound` to `Found` as the name of the sniff is always included in the errorcode anyway, so this was superfluous.
* As the name of the file being included/required will now be checked, added a safeguard against throwing the error too early during live coding + added a unit test for this.

Closes 104